refactor(test/helm): split portdrift_test.go into topical files
portdrift_test.go had grown to 1033 lines holding 11 top-level Test
functions but was named for only one of them (TestChartPortsMatchSvcinfo).
Split it into topically-named files in the same `helm` package so each
concern is findable by its filename, closing the "11 concerns under one
name" tech-debt.

This is a PURE MOVE: no test logic, assertion, helper behavior, or doc
comment changed. Verified by a byte-identical multiset diff of every
code+comment line (914 lines) between the original file and the union of
the new files.

Test functions now live in:
- portdrift_test.go          TestChartPortsMatchSvcinfo (the port-drift check)
- statestore_chart_test.go   TestStatestoreChartPorts, TestStateSvcChart
- workflow_chart_test.go     TestWorkflowChart
- asyncinvoke_chart_test.go  TestAsyncInvocationChart
- webhook_rbac_test.go       TestWebhookEnvReaderRBACScopesByTenancy
- buildermgr_chart_test.go   TestBuildermgrCanRollBuilderDeployments
- networkpolicy_test.go      TestNetworkPolicySvcLabelsResolveToRealWorkloads
- nameformat_test.go         TestNameFormat
- internalauth_chart_test.go TestInternalAuthExistingSecret, TestInternalAuthAutoGenerate

All package-shared, non-Test helpers (render, renderAs, renderErr,
helmTemplate, find, servicePorts, containerArgs, containerPorts,
containerEnv, argAfter, countByKindName, selectorSvcLabel, ingressSvcLabels,
npAllowsFromSvc, stringsOf, verbsFor, podSelectorSvcLabels, jobNamesFor,
jobNamesFrom, internalAuthEnvNames, envValues, podTemplates,
forEachContainerEnv) move to a single helmrender_test.go, so every split
file — and the pre-existing authvap/executorposture/clusterdomain tests in
the same package — compiles against exactly one definition of each.

go vet ./test/helm/ and go test ./test/helm/ both pass (14 tests, 0 skips).
